% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bgsig.R
\name{bootstrapGenomesfun2}
\alias{bootstrapGenomesfun2}
\title{Wrap for subtracting the background to estimate the experimentally generated signature}
\usage{
bootstrapGenomesfun2(genomes, n)
}
\arguments{
\item{genomes}{mutational catalog for bootstrapping}

\item{n}{number of bootstrapping to draw}
}
\value{
a matrix of bootstrapping sample catalogs
}
\description{
Wrap for subtracting the background to estimate the experimentally generated signature
}
