% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotCountbasis}
\alias{plotCountbasis}
\title{Plot mutational catalog}
\usage{
plotCountbasis(muts_basis, colnum, h, w, outputname)
}
\arguments{
\item{muts_basis}{96-channel mutational catalog}

\item{colnum}{number of columns in the plot}

\item{h}{height of plot}

\item{w}{width of plot}

\item{outputname}{name of output}
}
\description{
Plot mutational catalog
}
