% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bgsig_generic.R
\name{RemoveBackground_vector_single_channel}
\alias{RemoveBackground_vector_single_channel}
\title{Return single vector}
\usage{
RemoveBackground_vector_single_channel(
  background_profile,
  sig_profile,
  sampling_number,
  start_num,
  boundary,
  channel_number
)
}
\arguments{
\item{background_profile}{background mutational catalog}

\item{sig_profile}{experimentally generated mutational catalog}

\item{sampling_number}{number of bootstrapping samples}

\item{start_num}{mutation burden to start}

\item{boundary}{Range of signature around centroid (default = 2)}

\item{channel_number}{number of channels in user-defined catalogue}
}
\value{
data.frame including background signature and experiment signature
}
\description{
Return single vector
}
